refactor(base): rename to `TemplatePreview`
